Catch `Net::SCP::Error` exceptions and print messages to STDERR.
Sometimes, a directory may exist on some EY instances, but not all
of them. For example, a Rails app may be installed on app servers
and utility servers, but not the DB servers. Depending on the order
that `ey-scp` chooses to iterate through the server list, it might
try (and fail) to push a Rails configuration file (for example) to
a DB server before it has copied the file to all of the servers that
need it. Before, this would result in some of the servers receiving
the file, while others fail to receive it. Now, failure to copy a
file to one EY instance will not stop the gem from copying the file
to the other, remaining instances.